+++ src/main/java/com/enjoytrip/dto/user/UserGrade.java
@@ -1,4 +1,4 @@
-package com.enjoytrip.user.dto;
+package com.enjoytrip.dto.user;
 
 import java.sql.Timestamp;
